AccountServiceModel: explicitly delete private class
Leaving the deletion of the private class to the QObject child-parent
relationship means that the private class will be deleted only when the
AccountServiceModel object has already been destroyed. This means that
there is a window of time when the AccountServiceModelPrivate class
could attempt to access the already destroyed AccountServiceModel
object.
By explicitly deleting the AccountServiceModelPrivate class in the
AccountServiceModel destructor we make sure that such invalid accesses
can not happen.